utility functions for writing to local storage
